Update license GNU GPL.txt
The previous license was inappropriate and was updated to match the license used by the original program, egonet:  http://sourceforge.net/projects/egonet/ and the first version of EgoWeb: https://github.com/ericlavigne/egoweb.  As EgoWeb 2.0 is a derivative of these programs, it could not use the less restrictive MIT license.